// apps/backend/src/dataSource/types.ts
export interface IKaggleDataset {
  titleNullable: string;
  creatorNameNullable: string;
  totalBytesNullable: number;
  creatorUrlNullable: string;
  ownerNameNullable: string;
  ownerRefNullable: string;
  urlNullable: string;
  ref: string;
  creatorUrl: string;
  hasCreatorUrl: boolean;
  totalBytes: number;
  hasTotalBytes: boolean;
  url: string;
  hasUrl: boolean;
  ownerRef: string;
  hasOwnerRef: boolean;
}